backend/requirements.txt
fastapi
uvicorn
python-multipart
imagehash
Pillow
opencv-python-headless
numpy
pydantic
sqlalchemy
python-dotenv
google-generativeai